engine_configscan/engine_configscan_gcp/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
pyyaml>=6.0
python-dotenv>=0.21.0
httpx>=0.25.2
google-auth>=2.23.0
google-cloud-resource-manager>=1.12.0
